Implement common pinning API, include kWire Platform binaries

+++ kwire-runtime/src/x32Main/kotlin/dev/karmakrafts/kwire/memory/Memory.kt
@@ -89,7 +89,7 @@ private object NativeMemory : Memory {
     }
 
     override fun strlen(address: Address): NUInt {
-        return strlen_with_address(address.toCPointer()).toInt().toNUInt()
+        return strlen_with_address(address.toCPointer()).toNUInt()
     }
 
     override fun strcpy(source: Address, dest: Address) {
